added main program while loop, keyboard inputs
added if statements that check for keyboard inputs and change x or y variable by 3 to allow for movement of mysprite. Also added functionality of pressing space key to get mysprite back to the original position. Finally, added for loop that allows user to quite pygame window or press escape to exit the pygame window. -- 5 hrs.